+++ src/dropzone/scope-prefill.js
@@ -1,33 +1,35 @@
 import AbstractBuilder from "../abstract-builder";
-import { identity } from "../browser-utility";
 import TemplateElement from "../content-element/template-element";
-import DesignJsonPropertyExtension from "../design-json-property-extension";
-import RawValue from "../raw-value";
 
-/** @typedef {import('../content-element/content-element').default} ContentElement */
 /** @typedef {import('../content-element/template-element').default} TemplateElement */
 
 /**
  * This is the builder class to specify a scope prefill for a dropzone.
  *
- * @example
- * .withDropzones(
- *   cx.dropzone
- *     .withDropzone('a5142bca-448b-40c5-bdde-942f531fcd12')
- *     .withAllowedElements(
- *       require('./content-elements/basic/text'),
- *       require('./content-elements/basic/image'))
- *     .withMaxAllowedElements(1),
- *   cx.dropzone
- *     .withDropzone('3b369b8b-f1f6-4754-bb0f-e49a46c315e1')
- *     .withAllowedElements(
- *       require('./content-elements/basic/text'),
- *       require('./content-elements/basic/image'))
- *     .withMaxAllowedElements(1))
+ * @example cx.ScopePrefill('scopeA', require('./my-element'));
+ *
+ * use it within the Dropzone to define the prefill
+ *
+ * @example cx.Dropzone(..)
+ *   .withScopePrefills(cx.ScopePrefill('scopeA', require('./my-element')));
+ *
  */
 export default class ScopePrefill extends AbstractBuilder {
-  _scope = undefined;
-  _element = undefined;
+  /**
+   * @type {string}
+   * @private
+   */
+  _scope;
+  /**
+   * @type {TemplateElement}
+   * @private
+   */
+  _element;
+  /**
+   * @type {Object}
+   * @private
+   */
+  _overrideValues = {};
 
   constructor(scope, element) {
     super();
@@ -46,5 +48,44 @@ export default class ScopePrefill extends AbstractBuilder {
    */
   get element() {
     return this._element;
+  }
+  /**
+   * @returns {Object}
+   */
+  get overrideValues() {
+    return this._overrideValues;
+  }
+
+  /**
+   * Shorthand to overwrite prefill values within the element.
+   *
+   * @example
+   * cx.scopePrefill('scope', require('element'))
+   *   .withOverwriteValue('part-id', 'different-text')
+   *
+   * @param {string} templatePartId
+   * @param {string} value
+   * @returns {ScopePrefill}
+   */
+  withOverrideValue(templatePartId, value) {
+    this._overrideValues[templatePartId] = value;
+    return this;
+  }
+
+  /**
+   * Add scope with element to contextFile
+   * 
+   * @protected
+   * @param {Object} contextFile 
+   */
+  addPrefillTo(contextFile) {
+    this.element._loadPrefillIntoContextFile();
+    const context = this.element.contextFile;
+    let override = Object.entries(this.overrideValues);
+    override.forEach(
+      ([templatePartId, value]) =>
+        (context[templatePartId].value = value),
+    );
+    contextFile[this.scope] = context;
   }
 }